Очередной фикс варнингов

+++ trunk/xrGame/ai/monsters/control_animation.cpp
@@ -105,7 +105,7 @@ void CControlAnimation::play_part(SAnimationPart &part, PlayCallback callback)
 	// initialize synchronization of prev and current animation
 	float pos		= -1.f;
 	if (part.blend && !part.blend->stop_at_end) 
-		pos			= fmod(part.blend->timeCurrent,part.blend->timeTotal)/part.blend->timeTotal;
+		pos			= static_cast<float>(fmod(static_cast<double>(part.blend->timeCurrent),static_cast<double>(part.blend->timeTotal))/static_cast<double>(part.blend->timeTotal));
 #ifdef DEBUG
 	//CKinematicsAnimated * K = m_object->Visual()->dcast_PKinematicsAnimated();
 	//Msg				("%6d Playing animation : %s , %s , Object %s",Device.dwTimeGlobal, K->LL_MotionDefName_dbg(part.motion).first,K->LL_MotionDefName_dbg(part.motion).second, *(m_object->cName()));
