+++ trunk/xrGame/ik/eqn.cxx
@@ -64,7 +64,7 @@ static int solve_trig1_aux(float c,
     if (temp < 0.0f)
 	return 0;
 
-    temp  = atan2(_sqrt(temp), c);
+    temp  = static_cast<float>(atan2(static_cast<double>(_sqrt(temp)), static_cast<double>(c)));
     num =  (_abs(temp) > 1e-6f) ? 2 : 1;
 
     theta[0] = atan2ba;
@@ -97,7 +97,7 @@ static int solve_trig1_aux(float c,
 
 static int solve_trig1(float a, float b, float c, float theta[2])
 {
-    return solve_trig1_aux(c, a*a+b*b, atan2(b,a), theta);
+    return solve_trig1_aux(c, a*a+b*b, static_cast<float>(atan2(static_cast<double>(b),static_cast<double>(a))), theta);
 }
 
 #if 0
